passing all tests

+++ aot/src/main/java/com/dylibso/chicory/experimental/aot/AotAnalyzer.java
@@ -559,7 +559,7 @@ final class AotAnalyzer {
                 break;
             case REF_NULL:
                 // [] -> [ref]
-                stack.push(ValueType.refTypeForId((int) ins.operand(0)));
+                stack.push(ValueType.refTypeForId(ins.operand(0)));
                 break;
             case REF_IS_NULL:
                 // [ref] -> [I32]
@@ -692,14 +692,14 @@ final class AotAnalyzer {
     }
 
     private FunctionType blockType(Instruction ins) {
-        var typeId = (int) ins.operand(0);
+        var typeId = ins.operand(0);
         if (typeId == 0x40) {
             return FunctionType.empty();
         }
         if (ValueType.isValid(typeId)) {
             return FunctionType.returning(ValueType.forId(typeId));
         }
-        return module.typeSection().getType(typeId);
+        return module.typeSection().getType((int) typeId);
     }
 
     private static List<ValueType> getGlobalTypes(WasmModule module) {
